requirements.txt
# requirements.txt
fastapi==0.104.1
python-socketio==5.9.0
sqlalchemy==2.0.23
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1